+++ api/src/main/java/ca/bc/gov/educ/api/gradstudent/EducGradStudentApiApplication.java
@@ -1,31 +1,14 @@
 package ca.bc.gov.educ.api.gradstudent;
 
-import net.javacrumbs.shedlock.core.LockProvider;
-import net.javacrumbs.shedlock.provider.jdbctemplate.JdbcTemplateLockProvider;
 import net.javacrumbs.shedlock.spring.annotation.EnableSchedulerLock;
-import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.scheduling.annotation.EnableScheduling;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.context.SecurityContextHolder;
-
-import ca.bc.gov.educ.api.gradstudent.model.dto.GraduationStudentRecord;
-import ca.bc.gov.educ.api.gradstudent.model.entity.GraduationStudentRecordEntity;
-import org.springframework.transaction.PlatformTransactionManager;
 
 
 @SpringBootApplication
-@EnableGlobalMethodSecurity(prePostEnabled = true)
 @EnableCaching
 @EnableScheduling
 @EnableSchedulerLock(defaultLockAtMostFor = "1s")
@@ -36,54 +19,4 @@ public class EducGradStudentApiApplication {
         SpringApplication.run(EducGradStudentApiApplication.class, args);
     }
 
-    @Bean
-    public ModelMapper modelMapper() {
-
-        ModelMapper modelMapper = new ModelMapper();
-        modelMapper.typeMap(GraduationStudentRecordEntity.class, GraduationStudentRecord.class).addMappings(mapper -> mapper.skip(GraduationStudentRecord::setProgramCompletionDate));
-        modelMapper.typeMap(GraduationStudentRecord.class, GraduationStudentRecordEntity.class).addMappings(mapper -> mapper.skip(GraduationStudentRecordEntity::setProgramCompletionDate));
-        return modelMapper;
-    }
-
-
-    @Configuration
-    static
-    class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
-        /**
-         * Instantiates a new Web security configuration.
-         * This makes sure that security context is propagated to async threads as well.
-         */
-        public WebSecurityConfiguration() {
-            super();
-            SecurityContextHolder.setStrategyName(SecurityContextHolder.MODE_INHERITABLETHREADLOCAL);
-        }
-
-        @Override
-        public void configure(WebSecurity web) {
-            web.ignoring().antMatchers("/api/v1/api-docs-ui.html",
-                    "/api/v1/swagger-ui/**", "/api/v1/api-docs/**",
-                    "/actuator/health", "/actuator/prometheus", "/health");
-        }
-        @Override
-        protected void configure(final HttpSecurity http) throws Exception {
-            http.authorizeRequests()
-                    .anyRequest()
-                    .authenticated()
-                    .and()
-                    .oauth2ResourceServer()
-                    .jwt();
-        }
-    }
-
-    /**
-     * Lock provider lock provider.
-     *
-     * @param jdbcTemplate       the jdbc template
-     * @param transactionManager the transaction manager
-     * @return the lock provider
-     */
-    @Bean
-    public LockProvider lockProvider(@Autowired JdbcTemplate jdbcTemplate, @Autowired PlatformTransactionManager transactionManager) {
-        return new JdbcTemplateLockProvider(jdbcTemplate, transactionManager, "STATUS_SHEDLOCK");
-    }
 }
