+++ api/src/test/resources/application.yaml
@@ -5,7 +5,7 @@ spring:
   datasource:
     hikari:
       pool-name: gradstudent-api-cp
-      maximum-pool-size: 1
+      maximum-pool-size: 5
       connection-timeout: 6000
       max-lifetime: 600000
     driver-class-name: org.h2.Driver
